+++ js/technical/temp.js
@@ -23,8 +23,9 @@ function setupTemp() {
 		tmp[layer].resetGain = {}
 		tmp[layer].nextAt = {}
 		tmp[layer].nextAtDisp = {}
-		tmp[layer].notify = {}
 		tmp[layer].canReset = {}
+		tmp[layer].notify = {}
+		tmp[layer].prestigeNotify = {}
 		tmp[layer].prestigeButtonText = {}
 		setupBarStyles(layer)
 	}
@@ -63,8 +64,9 @@ function updateTemp() {
 		tmp[layer].resetGain = getResetGain(layer)
 		tmp[layer].nextAt = getNextAt(layer)
 		tmp[layer].nextAtDisp = getNextAt(layer, true)
-		tmp[layer].notify = shouldNotify(layer)
 		tmp[layer].canReset = canReset(layer)
+		tmp[layer].notify = shouldNotify(layer)
+		tmp[layer].prestigeNotify = prestigeNotify(layer)
 		tmp[layer].prestigeButtonText = prestigeButtonText(layer)
 		constructBarStyles(layer)
 	}